+++ sdk/core/azure-core/test/ut/pipeline.cpp
@@ -1,9 +1,9 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // SPDX-License-Identifier: MIT
 
-#include <gtest/gtest.h>
 #include <azure/core/http/pipeline.hpp>
 #include <azure/core/http/policy.hpp>
+#include <gtest/gtest.h>
 
 #include <vector>