+++ sdk/storage/azure-storage-common/src/shared_key_policy.cpp
@@ -18,18 +18,17 @@ namespace Azure { namespace Storage { namespace Details {
     string_to_sign += Azure::Core::Http::HttpMethodToString(request.GetMethod()) + "\n";
 
     const auto& headers = request.GetHeaders();
-    for (std::string headerName :
-         {"Content-Encoding",
-          "Content-Language",
-          "Content-Length",
-          "Content-MD5",
-          "Content-Type",
-          "Date",
-          "If-Modified-Since",
-          "If-Match",
-          "If-None-Match",
-          "If-Unmodified-Since",
-          "Range"})
+    for (std::string headerName : {"Content-Encoding",
+                                   "Content-Language",
+                                   "Content-Length",
+                                   "Content-MD5",
+                                   "Content-Type",
+                                   "Date",
+                                   "If-Modified-Since",
+                                   "If-Match",
+                                   "If-None-Match",
+                                   "If-Unmodified-Since",
+                                   "Range"})
     {
       auto ite = headers.find(Azure::Core::Strings::ToLower(headerName));
       if (ite != headers.end())