+++ sdk/storage/azure-storage-common/src/storage_per_retry_policy.cpp
@@ -2,6 +2,7 @@
 // SPDX-License-Identifier: MIT
 
 #include <azure/core/platform.hpp>
+
 #include "azure/storage/common/storage_per_retry_policy.hpp"
 
 #include <ctime>
@@ -23,7 +24,7 @@ namespace Azure { namespace Storage { namespace Details {
       // TODO: call helper function provided by Azure Core when they provide one.
       time_t t = std::time(nullptr);
       struct tm ct;
-#ifdef AZ_PLATFORM_WINDOWS
+#if defined(AZ_PLATFORM_WINDOWS)
       gmtime_s(&ct, &t);
 #elif defined(AZ_PLATFORM_POSIX)
       gmtime_r(&t, &ct);