+++ sdk/storage/azure-storage-files-shares/src/share_sas_builder.cpp
@@ -5,7 +5,7 @@
 #include "azure/core/http/http.hpp"
 #include "azure/storage/common/crypt.hpp"
 
-namespace Azure { namespace Storage { namespace Files { namespace Shares {
+namespace Azure { namespace Storage { namespace Sas {
 
   namespace {
     std::string ShareSasResourceToString(ShareSasResource resource)
@@ -80,66 +80,71 @@ namespace Azure { namespace Storage { namespace Files { namespace Shares {
     {
       canonicalName += "/" + FilePath;
     }
-    std::string protocol = Storage::Details::SasProtocolToString(Protocol);
+    std::string protocol = Details::SasProtocolToString(Protocol);
     std::string resource = ShareSasResourceToString(Resource);
 
     std::string stringToSign = Permissions + "\n" + (StartsOn.HasValue() ? StartsOn.GetValue() : "")
         + "\n" + ExpiresOn + "\n" + canonicalName + "\n" + Identifier + "\n"
         + (IPRange.HasValue() ? IPRange.GetValue() : "") + "\n" + protocol + "\n"
-        + Details::DefaultSasVersion + "\n" + CacheControl + "\n" + ContentDisposition + "\n"
-        + ContentEncoding + "\n" + ContentLanguage + "\n" + ContentType;
+        + Storage::Details::DefaultSasVersion + "\n" + CacheControl + "\n" + ContentDisposition
+        + "\n" + ContentEncoding + "\n" + ContentLanguage + "\n" + ContentType;
 
-    std::string signature
-        = Base64Encode(Details::HmacSha256(stringToSign, Base64Decode(credential.GetAccountKey())));
+    std::string signature = Base64Encode(
+        Storage::Details::HmacSha256(stringToSign, Base64Decode(credential.GetAccountKey())));
 
     Azure::Core::Http::Url builder;
     builder.AppendQueryParameter(
-        "sv", Details::UrlEncodeQueryParameter(Details::DefaultSasVersion));
-    builder.AppendQueryParameter("spr", Details::UrlEncodeQueryParameter(protocol));
+        "sv", Storage::Details::UrlEncodeQueryParameter(Storage::Details::DefaultSasVersion));
+    builder.AppendQueryParameter("spr", Storage::Details::UrlEncodeQueryParameter(protocol));
     if (StartsOn.HasValue())
     {
-      builder.AppendQueryParameter("st", Details::UrlEncodeQueryParameter(StartsOn.GetValue()));
+      builder.AppendQueryParameter(
+          "st", Storage::Details::UrlEncodeQueryParameter(StartsOn.GetValue()));
     }
     if (!ExpiresOn.empty())
     {
-      builder.AppendQueryParameter("se", Details::UrlEncodeQueryParameter(ExpiresOn));
+      builder.AppendQueryParameter("se", Storage::Details::UrlEncodeQueryParameter(ExpiresOn));
     }
     if (IPRange.HasValue())
     {
-      builder.AppendQueryParameter("sip", Details::UrlEncodeQueryParameter(IPRange.GetValue()));
+      builder.AppendQueryParameter(
+          "sip", Storage::Details::UrlEncodeQueryParameter(IPRange.GetValue()));
     }
     if (!Identifier.empty())
     {
-      builder.AppendQueryParameter("si", Details::UrlEncodeQueryParameter(Identifier));
+      builder.AppendQueryParameter("si", Storage::Details::UrlEncodeQueryParameter(Identifier));
     }
-    builder.AppendQueryParameter("sr", Details::UrlEncodeQueryParameter(resource));
+    builder.AppendQueryParameter("sr", Storage::Details::UrlEncodeQueryParameter(resource));
     if (!Permissions.empty())
     {
-      builder.AppendQueryParameter("sp", Details::UrlEncodeQueryParameter(Permissions));
+      builder.AppendQueryParameter("sp", Storage::Details::UrlEncodeQueryParameter(Permissions));
     }
-    builder.AppendQueryParameter("sig", Details::UrlEncodeQueryParameter(signature));
+    builder.AppendQueryParameter("sig", Storage::Details::UrlEncodeQueryParameter(signature));
     if (!CacheControl.empty())
     {
-      builder.AppendQueryParameter("rscc", Details::UrlEncodeQueryParameter(CacheControl));
+      builder.AppendQueryParameter("rscc", Storage::Details::UrlEncodeQueryParameter(CacheControl));
     }
     if (!ContentDisposition.empty())
     {
-      builder.AppendQueryParameter("rscd", Details::UrlEncodeQueryParameter(ContentDisposition));
+      builder.AppendQueryParameter(
+          "rscd", Storage::Details::UrlEncodeQueryParameter(ContentDisposition));
     }
     if (!ContentEncoding.empty())
     {
-      builder.AppendQueryParameter("rsce", Details::UrlEncodeQueryParameter(ContentEncoding));
+      builder.AppendQueryParameter(
+          "rsce", Storage::Details::UrlEncodeQueryParameter(ContentEncoding));
     }
     if (!ContentLanguage.empty())
     {
-      builder.AppendQueryParameter("rscl", Details::UrlEncodeQueryParameter(ContentLanguage));
+      builder.AppendQueryParameter(
+          "rscl", Storage::Details::UrlEncodeQueryParameter(ContentLanguage));
     }
     if (!ContentType.empty())
     {
-      builder.AppendQueryParameter("rsct", Details::UrlEncodeQueryParameter(ContentType));
+      builder.AppendQueryParameter("rsct", Storage::Details::UrlEncodeQueryParameter(ContentType));
     }
 
     return builder.GetAbsoluteUrl();
   }
 
-}}}} // namespace Azure::Storage::Files::Shares
+}}} // namespace Azure::Storage::Sas