+++ sdk/core/azure-core/CHANGELOG.md
@@ -5,6 +5,9 @@
 ### New Features
 
 - Added a WinHTTP-based `HttpTransport` called `WinHttpTransport` and use that as the default `TransportPolicyOptions.Transport` on Windows when sending and receiving requests and responses over the wire.
+- Added `Range` type to `Azure::Core::Http` namespace. 
+- Added support for long-running operations with `Operation<T>`.
+- Added support for setting a custom transport adapter by implementing the method `std::shared_ptr<HttpTransport> ::AzureSdkGetCustomHttpTransport()`.
 
 ### Breaking Changes
 
@@ -16,6 +19,7 @@
 ### Bug Fixes
 
 - Fixed for the Curl transport adapter when setting options.
+- Fixed for setting up the default transport adapter.
 
 ## 1.0.0-beta.3 (2020-11-11)