+++ sdk/storage/azure-storage-common/CHANGELOG.md
@@ -6,6 +6,7 @@
 - Rename `StorageSharedKeyCredential::UpdateAccountKey` to `StorageSharedKeyCredential::Update`.
 - Move `StorageRetryPolicy`, `StoragePerRetryPolicy` and `SharedKeyPolicy` to `Details` namespace.
 - Remove `StorageRetryOptions`, use `Azure::Core::Http::RetryOptions` instead.
+- Move Account SAS into `Azure::Storage::Sas` namespace.
 
 ## 12.0.0-beta.5 (2020-11-13)